requirements.txt
flask>=2.0.0
pandas>=2.0.0
mcp>=0.6.0
anthropic>=0.8.0
python-dotenv>=1.0.0
